Update string.h
strings are doing very well - _data is no longer const because it needs
to be manipulated. String constructors copy data so that variables that
aren't part of string are never manipulated. Added concat() function for
both other string objects and const char* - as well as operator += as a
substitute for concat() and operator + which returns a new string
concatenated to the argument.