various optimizations
many function params are now const refs, so code should work better
especially for large values. the copy function is also better now as it
returns a string and not a char* because otherwise to convert to string
first the char would have to be copied from the first string and then
copied again into the second string. now if you want a char* you just
have to do string.copy().text()