Various changes
First, exceptions were improved to the point where they can be caught
and are somewhat useful. However, the messages of exceptions that are
thrown are not particularly helpful.

String was improved by having the default constructor from a char* not
copy the value but rather simply use it. It can be copied with the
separate constructor string(char*,bool).

The add(T value) was added to the list class. This makes it possible to
add values to a list. However, this requires copying the entire list
over.

The get() method of the void class was changed to a reference. Testing
still needs to be done with this. It also catches bad_alloc exceptions
and throws a new exception with an unhelpful message.