Add list<T>.insert(T,int) method; consistencify
The insert(T value, int index) method inserts a new value right before
the specified index. Thus, list<string>.insert("Hello, world!",0) will
insert "Hello, world!" as the new first element of the list. The
operands in list.bubble(int,int) have also been switched around so
instead of specifying the index and then the amount of new elements, the
amount of elements and then the index is specified. This is so it has
the same order of operands as the insert() method.