Create byte and bitset.h
Byte and bitset are WIP headers which will be used for memory
efficiency. The byte is an unsigned char so it is essentially an
unsigned 8-bit integer. I'm not sure if it should be signed or unsigned.
It is hard to say. The BitSet is a class which contains a set of bits.
These bits can be set individually. The size of the BitSet class (in
bytes) is equal to (number of bits / 8) + 1. The BitString class is
similar to a BitSet but only includes 1 byte of bits.