Lots of varied changes
First: All .h files are replaced with .hpp to make it abundantly clear
that they are C++ code and are NOT compatible with C. The #pragma once
that is unique to VC++ was also abolished and the simple header guard is
now considered "good enough" for now. byte.h was replaced with memory.h,
which has a bunch of things dealing with memory. First is the idx type
used for size / indexing of lists / strings. Lists / strings were
updated to not be extendable past these maximum sizes. memory.h also is
now the container file for the bit enum, which was also changed to a
byte so it wouldn't take up any more space than necessary. The
getNullified<T>() and getNullifiedSet<T>(length) return single values
and sets of values that have been initialized nullified - in memory,
they are set to a set of zeroes which is equal in binary width to the
width / size / memory consumption / whatever of the type passed to it.
So for an int, which is 4 bytes, it initializes four bytes in memory set
to 00000000. However, memory also has the calculateByteLength() method,
which calculates the possible byte length by shifting a bit to the left
until it no longer exists. This should allow all functions using this
value to work on machines with byte lengths between 1 and 65536.
memory.h also contains the typeSizeInBits() function which simply
returns the size in bytes of an object multiplied by the amount of bits
in a byte. The BitSet class (and potentially the BitString class) appear
to be broken.